Split Go SDK into its own module under sdks/go/
Make sdks/go a standalone Go module (github.com/anaregdesign/lantern/sdks/go)
so external projects can depend on the client alone without pulling in
server/, cli/, or core/.

- Move generated proto code into the SDK module at sdks/go/gen/{graph,openapiv2}
  and update buf.gen.yaml accordingly.
- Add sdks/go/go.mod with only grpc/protobuf/grpc-gateway deps.
- Add root go.work + a 'replace ./sdks/go' directive so local builds keep
  working across both modules.
- Relocate bufconn-based integration tests out of sdks/go (which would
  otherwise re-introduce server/core deps) into tests/integration/ in the
  root module, and rewrite them to exercise the SDK through its public API.
- Wire CI and Dockerfile for the multi-module layout.
- Refresh README and AGENTS to document the new boundary and 'go get'
  instructions for external consumers.
