+++ buf.gen.yaml
@@ -7,22 +7,22 @@ managed:
     - module: buf.build/googleapis/googleapis
   override:
     - file_option: go_package_prefix
-      value: github.com/anaregdesign/lantern/gen/go
+      value: github.com/anaregdesign/lantern/sdks/go/gen
 plugins:
   - remote: buf.build/protocolbuffers/go
-    out: gen/go
+    out: sdks/go/gen
     opt:
       - paths=source_relative
   - remote: buf.build/grpc/go
-    out: gen/go
+    out: sdks/go/gen
     opt:
       - paths=source_relative
       - require_unimplemented_servers=false
   - remote: buf.build/grpc-ecosystem/gateway
-    out: gen/go
+    out: sdks/go/gen
     opt:
       - paths=source_relative
   - remote: buf.build/grpc-ecosystem/openapiv2
-    out: gen/openapiv2
+    out: sdks/go/gen/openapiv2
 inputs:
   - directory: proto
