+++ sdks/go/gen/openapiv2/graph/v1/graph.swagger.json
